chore: improve performance and move files around
