+++ src/components/NumberedTextArea/useLineNumbers.tsx
@@ -6,7 +6,7 @@ import {
   DEFAULT_LINE_NUMBER_TEXT_AREA_HEIGHT,
   convertLineNumbersToString,
 } from './utils';
-import { useScrollSync } from './useScrollSync';
+import { useScrollSync } from '../../hooks/useScrollSync';
 
 export const useLineNumbers = (
   lineNumberTextArea: React.RefObject<HTMLTextAreaElement>,
@@ -21,8 +21,19 @@ export const useLineNumbers = (
         : initialLineNumberState
     );
 
+    React.useEffect(() => {
+    if (contentTextArea?.current) {
+      setLineNumberTextAreaProps(
+        initialValue,
+        contentTextArea?.current?.scrollHeight
+      );
+    }
+  }, []);
+
+  // Sync the scrolling of the content line number textarea and content textarea
   useScrollSync(lineNumberTextArea, contentTextArea);
 
+  // Calculate new line numbers and line number textarea height and width (in cols)
   const setLineNumberTextAreaProps = (
     text?: string,
     scrollHeight: number = DEFAULT_LINE_NUMBER_TEXT_AREA_HEIGHT
@@ -33,19 +44,12 @@ export const useLineNumbers = (
     const value = convertLineNumbersToString(lineNumbers);
 
     setLineNumberTextAreaPropsState({
-      cols, 
+      cols,
       height: scrollHeight,
       value
     });
   };
 
-  React.useEffect(() => {
-    if (contentTextArea?.current) {
-      setLineNumberTextAreaProps(
-        initialValue,
-        contentTextArea?.current?.scrollHeight
-      );
-    }
-  }, []);
+
   return [lineNumberTextAreaProps, setLineNumberTextAreaProps] as const;
 };
